src/cli/requirements.txt
typer[all]==0.9.0
requests==2.31.0
configparser==6.0.0
PyYAML==6.0.1